test-requirements.txt
flake8
nose
pillow